Remove eager instance name loading
I've now refactored every place that needs names to get them
dynamically. Ideally, this should avoid paying the cost to load instance
names when they're not needed. And it avoids listing instances twice
most of the time, as in `chazz list`, where we already need to list the
instances anyway so we already have the names.